refactor(057): divide tests by behavior
